+++ NEWS.md
@@ -5,6 +5,7 @@
 * which.pmin_chr() include ties argument
 * new utils function noinf_min()
 * add_var(): debug when by is multiple columns
+* change parametrization of categorize function using .startdate, .stopdate.
 
 
 # stcsanalysis 0.1.4